Use specs to enrol instances in services
This mainly involved changing the schema in the data package, and some
consequent changes to how instances are detected and how stats are
exposed.

I rewrote the command line tool, taking a cue from the existing
structure (hanging things off the options as methods).

The listener's Sync could be more efficient when it sees an update to
a service, and probably in other circumstances.