Allow for >1 instance specifications in a service
This mainly involved changing the format in the data package, and some
consequent changes to how instances are detected.

I rewrote the command line tool, taking a cue from the existing
structure (hanging things off the options as methods).

The listener's Sync could be more efficient when it sees an update to
a service, and probably in other circumstances.