Group instances in the web UI rather than by selector
Allows the grouping for comparison to be given by the user interface;
from being due to the "selectors" (which decide which instances are
included), to being decided arbitrarily in the web UI. This makes it
easier to decide, post-hoc, that you're going to compare v3 with v4 of
something, for example.

Currently this just concatenates all the instance IDs into one big
query, which isn't going to be sustainable; but it's fine for this
stage.

Next step: actually allow that grouping to be changed in the UI (it's
hardwired to tag for the minute).

(The selectors are still distinguished in the stats (as "group"), so
could be used if required.)