Implement amberctl query
NB changed docker.io/{tag,image} to {tag,image}. Labels are supposed
(in Docker and Kubernetes) to be namespaced, so I think it is safe to
have these in apparent global scope. It also means we can refer to
them uniformaly everywhere (a difficulty if we are using selectors
over `Instance`s otherwise).