Add a bit of logging to agent for ease of troubleshooting
In particular, it's difficult to tell which IP the agent registers for
instances, and why.

This also adds the option of supplying the host IP address with an
environment entry, which is easier for using with docker-compose for
example.

(In the particular case I was looking at, because of a configuration
failure the agent was not getting a value for host IP, and therefore
looking it up by resolving `hostname`; however, since it was running
in a container, it got the container's IP, which is no use at all.)