Test the listener
The co-ordination is not all that easy, because the listener runs a
select loop; the `Listener.wait()` method is an attempt to let tests
synchronise on values from the other channels being processed. As long
as the channels (the docker events, and the store watcher) are
zero-size, they have to be dealt with before the wait value is
seen. Delicate!

This needs a local, synchronous implementation of a store, and some
abstraction of what the docker client is doing for us. There is a
modest abstraction of the docker client, but for now the docker client
library types are shown through.