Scripts to start the amber daemons and to run amberctl
The former is to run on each host, and the latter is a convenience if
you don't have the binary.

These are kept separate for now, because they may develop differently;
for example, we're not sure we want the daemons running in docker, so
it's best not to bake that in.