Run service updates through the template, use in prebake
At the minute this is limited to a single service, which is named by
the user at the outset. This makes the prebake simple (it can deal
with just a single service, and doesn't require configuration meddling
from the user), though obviously limited.

A more capable scheme would deal with all services, but most likely
require the user to edit configuration files or templates (at the
least, to specify how server blocks relate to upstream blocks).

The prebaked image generates an entire nginx configuration. To account
for nginx's operational quirks, there's a couple of failsafes:

 * the config must be in place before nginx starts (fail enough!), so
   there's a starting configuration which simply displays an
   "unavailable" message

 * the upstream block must not be empty (of primary servers); so, in
   the template, at least one server line is generated, pointing at an
   unavilable message via a unix socket. If there are instances, this
   is demoted to a "backup", meaning it will only be used if nginx
   decides the primary servers are all unavailable.