Change how the service name gets handled for the pre-baked image
If balagent takes the service name of interest as a command line image,
in order to support the pre-babed ngingx image, then it cannot be used
for other scenarios with a more elaborate nginx config that involved
multiple services.

So instead, expose the environment to the config template, so that it
can use the SERVICE env var to select which service it cares about.