Reform the store API
In part, the old API was informed by the mode of use of etcd. But it
was a bit unreliable (etcd does not return values in the same order
every time) and awkward to use -- most uses just put the values into a
map or slice then used that.

The old-style Foreach* methods are now available as procedures
instead.

 - prefer returning slices of things to iterators

 - distinguish between the details of a service `data.Service`, and
   what you get from the store, `store.ServiceInfo`, which may contain
   other fields depending on what you asked for.