Make webpack config a dependency of the web image
Previously the various `web/.something` files were baked into the
webbuild image, meaning it had to be rebuilt if they changed. However,
they are only needed for running webpack, so we don't want them to
trigger a rebuild of the `webbuild` image.

NB `web/package.json` is a dependency of the `webbuild` image, since
it is largely there for the purpose of `npm install`ing
dependencies. This is a lengthy install, so it's worth only running it
when necessary (that is after all the rationale for the build
images). It is *also* a dependency of the `web` image, since it
includes the script for running webpack.

(I changed the working directory for the `webbuild` image, to avoid
`/build/build` paths in the makefiles)