Keep track of local instances in agent listener
To avoid trying to remove instances that don't exist when a container
dies, keep track of which instances have been registered, and only
attempt to remove an instance if the container that died corresponds
to a local instance.

This also simplifies by removing the `reconcile` method, and just
treats reading in items (services, containers) like discovering new
ones. This does mean it's important to read the containers in first,
since reading in a service removes any instance that does not have a
corresponding container.

A more comprehensive overhaul might involve segregating each host's
instances in the store API, so that an agent only ever operates on its
own instances; and, when redefining a service, reading in the
instances then applying changes, rather than keeping track as it goes.