Use a <base> element to work locally and in gh-pages
Requirements:
 - http://localhost:4000/fluxctl works
 - http://weaveworks.github.io/flux/fluxctl works
 - (would be nice) http://.../fluxctl/ works

Using a <base> element means that any relative URLs are resolved
relative to the given URL, so we have to be careful to remember to
start URLs 'at the top'.

NB that the baseurl in _config.yml is for the gh-pages deployment, and
is overridden if you run the site with the makefile.