Loosen the association between go source files and exes
Previously, the Makefile knew which subdirectories contained go source
files for which executables.  That's not really sustainable in the face
of single executables that combine multiple components.  And relieving
the constraints on the source code layout will be nice.

So now, if any go source files have changed, the Makefile runs a single
"go install" command to build all executables, and then rebuilds the
appropriate containers based on which executables got updated.  Because
"go install" is incremental, this doesn't make much difference to build
times.  They might even get faster in some cases.