Keep track of viable instances, with a retry queue
If a connection fails, take it out of rotation. Use a heap to keep
track of which instances are due for a retry.

If a retry succeeds, promote the instance to active status; if it
fails, schedule the instance for another retry in the more distant
future.

Abandon an instance after a certain number of consecutive connection
attempts.

To revisit:

 - if there are no active instances, but there *are* retries queued,
   this will attempt to connect to the queued instances in turn; and
   if they all fail, it will keep trying.

 - an abandoned instance (that failed too many connection attempts in
   a row) is forgotten about; an update that mentions it again will
   return it to active status.