Handle the case where the docker client closes the event channel
I encountered a nil pointer dereference in readContainerIDs when
starting fluxd, due to a nil event from the go-dockerclient.  It's
unclear what the cause was: go-dockerclient will close the channel if
the docker connection fails, but it also has some race conditions in
that area.  Both seem somewhat unlikely in this case, and I've been
unable to reproduce the problem.  So tolerate it in future by simply
restarting the docker listener.