Use a single field for service address
A service either has both an IP address and port, or neither. To avoid
confusion, make it a single field that can have a value or nil.

This also adjusts the rules for when the instance port is displayed:
if it's been set either implicitly (in the service) address or
explicitly by the user, it's shown.