Use netutil.IPPort for service addresses, rather than net.TCPAddr
First, because service addresses are not specifically TCP
addresses.  They could be UDP addresses in the future.

Second, because we don't use them as a net.TCPAddr anywhere (one or two
places might arise is the balancer code is converted, but a simple
converter method will suffice there).

Third, TCPAddr has an IPv6-related Zone field, which has an impact on
the code even though we never care about it.