Make the rule name optional in fluxctl select
I don't like asking the user to supply a meaningless value.

The choice of encoding is sensitive: there is a tension between having
strings that don't look too out of place, and keeping the resulting
names short given the amount of randomness to encode. I've picked what
I think is a reasonable balance.