Separate liveness and Host records
This rejigs heartbeating so that it becomes a way of limiting the
lifetime of records. The motivation is to give hosts and instances a
lifetime that corresponds to the fluxd process.

This lifetime is implicit; the store.Store implementation is expected
to enforce it. In the case of `etcdStore`, each run is given a session
ID, which is maintained by the heartbeat, and added to e.g, Host
records. Any Host records that don't correspond to a live session ID
are discarded.

One consequence is that if you are registering a host, you need to
heartbeat at least once to establish a lifetime for the record. (This
is reflected in changes to the tests, e.g.)