Add methods for tidying up values in etcdstore
Session-scoped values (hosts and instances) in the etcdstore that
don't correspond to a live session are ignored, which is enough for
correctness.

However, they will still be in the database, and it would be good to
clean them out from time to time. To this end, I've introduced a
method that will remove old items (using CAD operations so it doesn't
tread on toes). The idea will be to run this periodically in the
daemon.

*NB* I made the various bits of etcd path elements into constants, and
 in doing so changed their values to keep the code honest.