Make the store a component
Because the store needs a heartbeater process, and will need a
clean-up process, it is more natural for it to exist as a component
that is run rather than just a passive set of operations.

I have kept the store.Store as-is for now and added an interface for
the "runnable" store component, to distinguish between maintaining
definitions (e.g., services), and telling the system about runtime
events (instances coming and going). So that, for example, `fluxctl`
can just depend on an interface for changing definitions, and querying
for instances, while `fluxd` actively maintains the records of
instances.