Rework InstancePool
Never abandon an instance - it's not clear why that is a better strategy
than simply increasing the delay between retries.

When an instance has failed and is being retried, treat it as tentative:
when a thread attempts to connect to it, remove it from the ready set so
that other threads won't use it until a connection succeeds or another
retry attempt is due.