Set host record after store is started
When starting, the agent adds a record showing that the host is 'up',
which will time out if the agent dies. In the etcdstore implementation
this depends on the heartbeater running to be considered "live". There
is also a process running that removes records that don't correspond
to a live heartbeat. Both the heartbeat and the remover run
immediately, then on a ticker.

There is thus a race between the heartbeat starting and making the
agent's host record live, and the remover running and removing
it. Logically, the agent needs to wait for the heartbeater to have
started before it puts the record in, so it will be considered live
straight away.

This commit introduces a way to wait for the store to start; it
becomes part of an interface in `store` so that mock implementations
can implement it.