Fix intermittently failing test in balagent_test.go
The template package docs say:

    If the value is a map and the keys are of basic type with a defined
    order ("comparable"), the elements will be visited in sorted key
    order.

netutil.IPPort is "comparable", according to the definition of that term
in the Go Spec.  But it is not "ordered" (because it is a struct), which
is what that sentence really describes.  So the iteration of the ingress
instance map ends up being randomly shuffled, and we need an explicit
sort to make the test reliable.